chore: move `useFetchSourceStudiesSourceDatasets` to general hooks folder (#467)
